requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
groq>=0.11.0
PyGithub>=2.3.0
httpx>=0.27.0
python-dotenv>=1.0.1
slack-sdk>=3.33.0
